Simplify remote spec fetching and auth header resolution
- Extract shared HttpFetch utility for all remote HTTP requests
- Extract AuthHeaderResolver for env-var, placeholder, and !cmd resolution
- Simplify ApiFileLoader to delegate fetch/resolution
- Remove duplicated cache from AuthJsonLoader
- Remove jsonLoader from SourceApi public surface; flow through SourceApi.create
- Drop redundant jsonLoader branch in YamlUtils.parseOpenApi
- Split pure resolution tests into AuthHeaderResolverTest
